rename: 테스트 컨테이너 패키지 이동 및 테스트 클래스 파일명 변경
