--- src/test/java/kr/hhplus/be/server/application/user/UserFacadeTest.java
@@ -1,56 +1,0 @@
-package kr.hhplus.be.server.application.user;
-
-import kr.hhplus.be.server.DatabaseClearExtension;
-import kr.hhplus.be.server.application.user.dto.UserCriteria;
-import kr.hhplus.be.server.application.user.dto.UserResult;
-import kr.hhplus.be.server.domain.balance.entity.Balance;
-import kr.hhplus.be.server.domain.balance.repository.BalanceRepository;
-import kr.hhplus.be.server.domain.user.entity.User;
-import kr.hhplus.be.server.domain.user.repository.UserRepository;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-@SpringBootTest
-@ExtendWith(DatabaseClearExtension.class)
-@ActiveProfiles("test")
-@DisplayName("[통합테스트] UserFacade")
-public class UserFacadeTest {
-
-    @Autowired
-    private UserFacade userFacade;
-
-    @Autowired
-    private UserRepository userRepository;
-
-    @Autowired
-    private BalanceRepository balanceRepository;
-
-    @Test
-    @DisplayName("[성공] 사용자 생성 - 잔고도 같이 생성")
-    void createUser_ShouldCreateUserAndBalance() {
-
-        // Arrange
-        UserCriteria.Create criteria = new UserCriteria.Create("추경현");
-
-        // Act
-        UserResult.Create result = userFacade.createUser(criteria);
-
-        // Assert
-        assertThat(result.id()).isEqualTo(1L);
-        assertThat(result.name()).isEqualTo("추경현");
-        assertThat(result.balance()).isEqualTo(0);
-
-        User user = userRepository.findById(result.id());
-        assertThat(user.getName()).isEqualTo("추경현");
-
-        Balance balance = balanceRepository.findByUserId(result.id());
-        assertThat(balance.getBalance()).isEqualTo(0);
-    }
-
-}
