+++ src/test/java/kr/hhplus/be/server/surpport/container/MysqlTestcontainersConfiguration.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server;
+package kr.hhplus.be.server.surpport.container;
 
 import jakarta.annotation.PreDestroy;
 import org.springframework.context.annotation.Configuration;
@@ -6,16 +6,17 @@ import org.testcontainers.containers.MySQLContainer;
 import org.testcontainers.utility.DockerImageName;
 
 @Configuration
-class TestcontainersConfiguration {
+class MysqlTestcontainersConfiguration {
 
 	public static final MySQLContainer<?> MYSQL_CONTAINER;
 
 	static {
-		MYSQL_CONTAINER = new MySQLContainer<>(DockerImageName.parse("mysql:8.0"))
-			.withDatabaseName("hhplus")
-			.withUsername("test")
-			.withPassword("test");
-		MYSQL_CONTAINER.start();
+        MYSQL_CONTAINER = new MySQLContainer<>(DockerImageName.parse("mysql:8.0"))
+                .withDatabaseName("hhplus")
+                .withUsername("test")
+                .withPassword("test");
+
+        MYSQL_CONTAINER.start();
 
 		System.setProperty("spring.datasource.url", MYSQL_CONTAINER.getJdbcUrl() + "?characterEncoding=UTF-8&serverTimezone=UTC");
 		System.setProperty("spring.datasource.username", MYSQL_CONTAINER.getUsername());
