+++ src/test/java/kr/hhplus/be/server/surpport/database/DatabaseClearExtension.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server;
+package kr.hhplus.be.server.surpport.database;
 
 import org.junit.jupiter.api.extension.BeforeEachCallback;
 import org.junit.jupiter.api.extension.ExtensionContext;
